Fix typo in minor version regexp.
The previous regular expression fragment `"\\.([0-9])+\\."` was
capturing each digit of the minor version individually which is
resulting in generated CMake versions of 1.0.0 for Poco 1.10.0.

Moving the + into the capture allows for multi-digit minor versions.
I checked the other lines and neither had this typo so we should be good
for major and patch versions already.